endpoint/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
python-dotenv>=1.0.0
supabase>=1.0.3
requests>=2.28.0
